refactoring

+++ src/forex/load/PriceM1.java
@@ -4,10 +4,10 @@ import java.util.Date;
 
 public class PriceM1 {
     private Date m1DateValue;  //Массив дат
-    private Double m1MinPrice;  //Массив минимума
-    private Double m1MaxPrice;  //Массив максимума
+    private float m1MinPrice;  //Массив минимума
+    private float m1MaxPrice;  //Массив максимума
 
-    public PriceM1(Date parsingDate, Double valueOf, Double valueOf1) {
+    public PriceM1(Date parsingDate, float valueOf, float valueOf1) {
         m1DateValue = parsingDate;
         m1MaxPrice = valueOf;
         m1MinPrice = valueOf1;
@@ -30,19 +30,19 @@ public class PriceM1 {
         this.m1DateValue = m1DateValue;
     }
 
-    public Double getM1MinPrice() {
+    public float getM1MinPrice() {
         return m1MinPrice;
     }
 
-    public void setM1MinPrice(Double m1MinPrice) {
+    public void setM1MinPrice(float m1MinPrice) {
         this.m1MinPrice = m1MinPrice;
     }
 
-    public Double getM1MaxPrice() {
+    public float getM1MaxPrice() {
         return m1MaxPrice;
     }
 
-    public void setM1MaxPrice(Double m1MaxPrice) {
+    public void setM1MaxPrice(float m1MaxPrice) {
         this.m1MaxPrice = m1MaxPrice;
     }
 }
